Relocate launcher to correct namespace and allow future remote launcher
